fixed(layout)clear styled-component
