+++ package.json
@@ -12,13 +12,16 @@
     "@types/react": "^17.0.37",
     "@types/react-bootstrap": "^0.32.28",
     "@types/react-dom": "^17.0.11",
+    "@types/react-router-dom": "^5.3.2",
     "@types/styled-components": "^5.1.18",
     "bootstrap": "^5.1.3",
     "react": "^17.0.2",
     "react-bootstrap-validation": "^0.1.11",
     "react-dom": "^17.0.2",
+    "react-router-dom": "^6.2.1",
     "react-scripts": "5.0.0",
     "styled-components": "^5.3.3",
+    "styled-reset": "^4.3.4",
     "typescript": "^4.5.4",
     "web-vitals": "^2.1.2"
   },
