--- src/components/Logo/Logo.tsx
@@ -1,9 +1,0 @@
-import React from "react";
-
-import logo from "../../assets/Logo.png";
-
-const Logo = () => {
-  return <img className="logo" src={`${logo}`} alt="logo" />;
-};
-
-export default Logo;
