+++ src/components/Ticket.tsx
@@ -1,5 +1,14 @@
-import React from "react";
-import styled from "styled-components";
+import React from 'react';
+
+import {
+  Card,
+  Header,
+  Price,
+  ColumnInfo,
+  ColumnInfoTitle,
+  ColumnInfoTime,
+  Content,
+} from '../styles/Ticket';
 
 export interface ITicket {
   ticket: {
@@ -41,7 +50,7 @@ const Ticket: React.FC<ITicket> = ({ ticket }) => {
             </ColumnInfo>
             <ColumnInfo>
               <ColumnInfoTitle>
-                Пересадки{/**TODO: add count of transplantation */}{" "}
+                Пересадки{/**TODO: add count of transplantation */}{' '}
               </ColumnInfoTitle>
               <span>hkg, jnb</span>
             </ColumnInfo>
@@ -52,45 +61,4 @@ const Ticket: React.FC<ITicket> = ({ ticket }) => {
   );
 };
 
-const Card = styled.div`
-  background-color: #fff;
-  height: max-content;
-  border-radius: 10px;
-  padding: 25px 50px 25px 20px;
-  text-transform: uppercase;
-  font-size: 12px;
-  margin-bottom: 20px;
-`;
-
-const Header = styled.div`
-  display: grid;
-  grid-template-columns: 2fr 1fr;
-  align-items: center;
-  font-size: 12px;
-  margin-bottom: 10px;
-`;
-
-const Content = styled.section`
-  display: grid;
-  grid-template-columns: repeat(3, 1fr);
-`;
-
-const ColumnInfo = styled.p`
-  > span {
-    display: block;
-  }
-`;
-
-const ColumnInfoTitle = styled.span`
-  color: #a0b0b9;
-`;
-
-const ColumnInfoTime = styled.span`
-  text-transform: lowercase;
-`;
-
-const Price = styled.span`
-  font-size: 18px;
-  color: #2196f3;
-`;
 export default Ticket;
