+++ src/components/index.ts
@@ -1,0 +1,7 @@
+import Filter from './Filter';
+import Logo from './Logo';
+import NumberOfTransplants from './NumberOfTransplants';
+import Ticket from './Ticket';
+import TicketBoard from './TicketBoard';
+
+export { Filter, Logo, NumberOfTransplants, Ticket, TicketBoard };
