+++ src/index.tsx
@@ -1,12 +1,10 @@
-import React from "react";
-import ReactDOM from "react-dom";
-import "./styles/reset.css";
-import "./styles/index.css";
-import App from "./App";
+import React from 'react';
+import ReactDOM from 'react-dom';
+import App from './App';
 
 ReactDOM.render(
   <React.StrictMode>
     <App />
   </React.StrictMode>,
-  document.getElementById("root")
+  document.getElementById('root')
 );
