+++ src/styles/Logo.tsx
@@ -1,0 +1,10 @@
+import styled from 'styled-components';
+
+export const NavLink = styled.a`
+  margin: 50px auto;
+  width: 20%;
+  img {
+    max-width: 150px;
+    height: 100%;
+  }
+`;
